feat: 实现重复报名弹窗提示, 以及用户活动列表与活动模块的相互跳转行为

+++ pkg/events/models/userEvent.go
@@ -1,7 +1,7 @@
 package models
 
 type UserEvent struct {
-	UserID  int  `gorm:"primaryKey"`
-	EventID int  `gorm:"primaryKey"`
-	Checked bool `gorm:"default:false"`
+	UserID  int64 `gorm:"primaryKey"`
+	EventID int64 `gorm:"primaryKey"`
+	Checked bool  `gorm:"default:false"`
 }
